master limiter lookahead: finish + review fixes; widen the timeline to Double
Two changes that ended up sharing files, so they land together.

## 1. Master limiter — Phases 3–5, the dual release, and two review rounds

Phase 3 (wire + DSL): `MasterStageDsl.Limiter` gains `lookaheadSeconds` and
KlangScript gains `.lookahead()`. Defaults to 0 on purpose — an authored limiter
is per-playback, so latency there would delay that playback against every other
one. Only the house limiter, which runs once on the summed mix, can afford it.
`MasterChain.limiters` is now internal so specs can assert the wire→DSP hop.

Phase 5 (latency): `MasterStage.latencyFrames`/`latencyMs` are reported into
`outputLatencyMs`, so the code- and block-highlight overlays stay aligned; the
offline renderer runs that many frames past its musical end so the tail is not
left in the delay ring.

Dual program-dependent release on the lookahead path. Kept for the loudness, NOT
the pumping: at an identical peak the mean level goes −1.59 → −0.55 dB. The
by-ear verdict on the pump itself was "really subtle", on deliberately unkind
material — that is recorded in the code so nobody reverts this for the wrong
reason.

Defects found by review and fixed:
  - `sampleCounter` overflowed Int after ~12.4 h, turning the deque eviction into
    an infinite loop inside the render callback
  - the min-hold ring was one slot short; the emptiness guard added for wrap
    safety is what activated it, discarding the whole hold window
  - `resizeSmoothing` allocated and corrupted state; `primeBoxes` seeded unity,
    worth +15.94 dBFS mid-duck
  - the detector guarded NaN but not ±Inf, which muted the mix for ~450 ms

## 2. Absolute frames Int → Double

`cursorFrame` advances every block for the life of the backend, playing or not.
As an Int it overflowed after 12.4 h at 48 kHz and audio silently stopped — no
crash, no error. A tab left open overnight was enough. Measured before the fix:
0/60 blocks produced audio near the boundary vs 55/60 at frame 0, and the
degradation starts *before* the wrap because scheduling round-trips through
seconds.

Double, not Long: it needs no exception to the "no Long in audio paths" rule, it
is a native JS number where Long is emulated and allocates, and it is exact for
integers to 2^53 (~5,950 years at 48 kHz).

The rule this establishes: **absolute timeline = Double, relative position =
Int**, converted once per block per voice. NB `IgniteContext.gateEndFrame` is
voice-RELATIVE and stays Int despite sharing a name with the absolute
`BlockContext.gateEndFrame`.

## Guards

All new specs mutation-checked. `LimiterLookaheadSpec` was written before the fix
and watched failing. `LongRunningTimelineSpec` covers a class invisible to every
other test, since nobody runs a suite for twelve hours.

Also: the Geraint Luff / Signalsmith credit for the lookahead construction, on
both CREDITS.MD and the in-app page; and the doc surfaces that described the old
chain order, the old attack value, or lookahead as "deferred".

Whole-project `./gradlew build` green, plus JS suites and the song benchmark.

+++ audio_be/src/commonMain/kotlin/KlangAudioRenderer.kt
@@ -55,7 +55,10 @@ class KlangAudioRenderer private constructor(
         master.reset()
     }
 
-    fun renderBlock(cursorFrame: Int, out: ShortArray) {
+    // NB `cursorFrame` is Double, not Int: it is an ABSOLUTE frame on the backend timeline, which
+    // grows for the life of the backend and overflows Int after ~12.4 h. Exact below 2^53
+    // (~5,950 years at 48 kHz). See RenderClock.cursorFrame. Per-sample offsets stay Int.
+    fun renderBlock(cursorFrame: Double, out: ShortArray) {
         clock.cursorFrame = cursorFrame
         mix.clear()
         engine.renderInto(mix, cursorFrame)
